fix search min chars fix undefined value on SearchNumberBox
